+++ CHANGELOG.md
@@ -3,6 +3,10 @@ Add setting to not show terrain when dragging a token
 
 The original Terrain Layer does not play nice with the Enhanced Terrain Layer.  Added code to make sure they can exist at the same time.
 
+Fixed issue with opacity
+
+Added different border for hidden terrain instead of setting the opacity
+
 #Version 1.0.17
 Added function to try and copy old data from TerrainLayer